[#179] refactor: 스케쥴러를 별도의 유스케이스로 분리
