+++ src/main/java/com/kiero/schedule/adapter/in/scheduler/DailyScheduleJob.java
@@ -3,7 +3,7 @@ package com.kiero.schedule.adapter.in.scheduler;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import com.kiero.schedule.application.service.ScheduleCommandService;
+import com.kiero.schedule.application.port.in.ScheduleSchedulerUseCase;
 
 import lombok.RequiredArgsConstructor;
 
@@ -11,11 +11,11 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class DailyScheduleJob {
 
-	private final ScheduleCommandService scheduleCommandService;
+	private final ScheduleSchedulerUseCase scheduleSchedulerUseCase;
 
 	@Scheduled(cron = "0 0 0 * * *", zone = "Asia/Seoul")
 	public void runDailyJob() {
-		scheduleCommandService.createTodayScheduleDetail();
+		scheduleSchedulerUseCase.createTodayScheduleDetail();
 
 	}
 }
