+++ src/main/java/com/kiero/schedule/application/port/in/ScheduleSchedulerUseCase.java
@@ -1,0 +1,7 @@
+package com.kiero.schedule.application.port.in;
+
+public interface ScheduleSchedulerUseCase {
+
+	void createTodayScheduleDetail();
+
+}
